requirements.txt
pyTelegramBotAPI==4.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
anthropic==0.18.1
google-genai
openpyxl==3.1.2
beautifulsoup4==4.12.3
requests==2.31.0
psutil
Flask==3.0.0
